AIIncidentIntelligencePlatform/notification/requirements.txt
kafka-python>=2.0.2
httpx>=0.24.0
python-dotenv>=1.0.0
